requirements.txt
fake-useragent
beautifulsoup4
requests
Pillow